Removed the local repository requirements for the build-util modules.

+++ plc4go/internal/plc4go/knxnetip/readwrite/model/KnxDatapoint.go
@@ -936,7 +936,7 @@ func KnxDatapointParse(io *utils.ReadBuffer, datapointType KnxDatapointType) (ap
 	case datapointType == KnxDatapointType_DPT_Char_ASCII: // STRING
 
 		// Simple Field (value)
-		value, _valueErr := io.ReadString(8)
+		value, _valueErr := io.ReadString(uint32((8)))
 		if _valueErr != nil {
 			return nil, errors.Wrap(_valueErr, "Error parsing 'value' field")
 		}
@@ -944,7 +944,7 @@ func KnxDatapointParse(io *utils.ReadBuffer, datapointType KnxDatapointType) (ap
 	case datapointType == KnxDatapointType_DPT_Char_8859_1: // STRING
 
 		// Simple Field (value)
-		value, _valueErr := io.ReadString(8)
+		value, _valueErr := io.ReadString(uint32((8)))
 		if _valueErr != nil {
 			return nil, errors.Wrap(_valueErr, "Error parsing 'value' field")
 		}
@@ -2381,7 +2381,7 @@ func KnxDatapointParse(io *utils.ReadBuffer, datapointType KnxDatapointType) (ap
 	case datapointType == KnxDatapointType_DPT_String_ASCII: // STRING
 
 		// Simple Field (value)
-		value, _valueErr := io.ReadString(112)
+		value, _valueErr := io.ReadString(uint32((112)))
 		if _valueErr != nil {
 			return nil, errors.Wrap(_valueErr, "Error parsing 'value' field")
 		}
@@ -2389,7 +2389,7 @@ func KnxDatapointParse(io *utils.ReadBuffer, datapointType KnxDatapointType) (ap
 	case datapointType == KnxDatapointType_DPT_String_8859_1: // STRING
 
 		// Simple Field (value)
-		value, _valueErr := io.ReadString(112)
+		value, _valueErr := io.ReadString(uint32((112)))
 		if _valueErr != nil {
 			return nil, errors.Wrap(_valueErr, "Error parsing 'value' field")
 		}
@@ -4911,7 +4911,7 @@ func KnxDatapointParse(io *utils.ReadBuffer, datapointType KnxDatapointType) (ap
 	case datapointType == KnxDatapointType_DPT_LanguageCodeAlpha2_ASCII: // STRING
 
 		// Simple Field (value)
-		value, _valueErr := io.ReadString(16)
+		value, _valueErr := io.ReadString(uint32((16)))
 		if _valueErr != nil {
 			return nil, errors.Wrap(_valueErr, "Error parsing 'value' field")
 		}
@@ -6551,13 +6551,13 @@ func KnxDatapointSerialize(io *utils.WriteBuffer, value api.PlcValue, datapointT
 	case datapointType == KnxDatapointType_DPT_Char_ASCII: // STRING
 
 		// Simple Field (value)
-		if _err := io.WriteString(8, "ASCII", value.GetString()); _err != nil {
+		if _err := io.WriteString(uint8((8)), "ASCII", value.GetString()); _err != nil {
 			return errors.Wrap(_err, "Error serializing 'value' field")
 		}
 	case datapointType == KnxDatapointType_DPT_Char_8859_1: // STRING
 
 		// Simple Field (value)
-		if _err := io.WriteString(8, "ISO-8859-1", value.GetString()); _err != nil {
+		if _err := io.WriteString(uint8((8)), "ISO-8859-1", value.GetString()); _err != nil {
 			return errors.Wrap(_err, "Error serializing 'value' field")
 		}
 	case datapointType == KnxDatapointType_DPT_Scaling: // USINT
@@ -7630,13 +7630,13 @@ func KnxDatapointSerialize(io *utils.WriteBuffer, value api.PlcValue, datapointT
 	case datapointType == KnxDatapointType_DPT_String_ASCII: // STRING
 
 		// Simple Field (value)
-		if _err := io.WriteString(112, "ASCII", value.GetString()); _err != nil {
+		if _err := io.WriteString(uint8((112)), "ASCII", value.GetString()); _err != nil {
 			return errors.Wrap(_err, "Error serializing 'value' field")
 		}
 	case datapointType == KnxDatapointType_DPT_String_8859_1: // STRING
 
 		// Simple Field (value)
-		if _err := io.WriteString(112, "ISO-8859-1", value.GetString()); _err != nil {
+		if _err := io.WriteString(uint8((112)), "ISO-8859-1", value.GetString()); _err != nil {
 			return errors.Wrap(_err, "Error serializing 'value' field")
 		}
 	case datapointType == KnxDatapointType_DPT_SceneNumber: // USINT
@@ -9460,7 +9460,7 @@ func KnxDatapointSerialize(io *utils.WriteBuffer, value api.PlcValue, datapointT
 	case datapointType == KnxDatapointType_DPT_LanguageCodeAlpha2_ASCII: // STRING
 
 		// Simple Field (value)
-		if _err := io.WriteString(16, "ASCII", value.GetString()); _err != nil {
+		if _err := io.WriteString(uint8((16)), "ASCII", value.GetString()); _err != nil {
 			return errors.Wrap(_err, "Error serializing 'value' field")
 		}
 	case datapointType == KnxDatapointType_DPT_Tariff_ActiveEnergy: // Struct
